test(spec): an empty round stream is refused only when the arm ran, and the recorder gets its own file
