+++ crates/rmlx-models/tests/spec_greedy_equivalence.rs
@@ -153,6 +153,10 @@ use std::path::Path;
 
 mod common;
 
+use common::round_stream::{
+    round_events, CapturedEvent, RoundStreamRecorder, EAGLE3_STEP_SWITCH_TARGET, ROUND_EVENT_FIELDS,
+};
+
 use rmlx_mlx::Device;
 use rmlx_models::arch;
 use rmlx_models::speculative::dflash::{dflash_generate, DFlashDrafter};
@@ -2438,7 +2442,7 @@ impl Loaded {
         Vec<u32>,
         Vec<f32>,
         Vec<DecidedBy>,
-        Vec<common::CapturedEvent>,
+        Vec<CapturedEvent>,
     ) {
         let ids = prompt.ids(&self.tokenizer);
         let block = self.round_block();
@@ -2476,7 +2480,7 @@ impl Loaded {
         ids: &[u32],
         block: usize,
         device: Device,
-    ) -> (usize, Vec<u32>, Vec<DecidedBy>, Vec<common::CapturedEvent>) {
+    ) -> (usize, Vec<u32>, Vec<DecidedBy>, Vec<CapturedEvent>) {
         let mut spec_ids: Vec<u32> = Vec::new();
         let mut decided_by: Vec<DecidedBy> = Vec::new();
         // EAGLE-3 is the one loop that consults the second switch, and its
@@ -2492,7 +2496,7 @@ impl Loaded {
                 ..
             }
         );
-        let recorder = common::RoundStreamRecorder::new();
+        let recorder = RoundStreamRecorder::new();
         let ran = tracing::subscriber::with_default(std::sync::Arc::clone(&recorder), || {
             let mut step = |s: &rmlx_models::ProbeStep| {
                 spec_ids.push(s.token_id);
@@ -2651,7 +2655,7 @@ impl Loaded {
         let mut switches: Vec<&str> = questions
             .iter()
             .filter(|(target, level, _)| {
-                common::BEHAVIOUR_SWITCH_TARGETS.contains(&target.as_str())
+                common::round_stream::BEHAVIOUR_SWITCH_TARGETS.contains(&target.as_str())
                     && *level == tracing::Level::TRACE
             })
             .map(|(target, ..)| target.as_str())
@@ -2660,17 +2664,12 @@ impl Loaded {
         switches.dedup();
         eprintln!("switches declined: {switches:?}");
         assert!(
-            !asks_step_switch || switches.contains(&common::EAGLE3_STEP_SWITCH_TARGET),
-            "this loop gates its per-position trace on {} and the capture was never \
-             asked about it, so the gate is no longer there: {switches:?}",
-            common::EAGLE3_STEP_SWITCH_TARGET
+            !asks_step_switch || switches.contains(&EAGLE3_STEP_SWITCH_TARGET),
+            "this loop gates its per-position trace on {EAGLE3_STEP_SWITCH_TARGET} and \
+             the capture was never asked about it, so the gate is no longer there: \
+             {switches:?}"
         );
-        (
-            ran,
-            spec_ids,
-            decided_by,
-            common::round_events(&recorder.events()),
-        )
+        (ran, spec_ids, decided_by, round_events(&recorder.events()))
     }
 
     /// The target-vocabulary ids this pair's drafter can name, or `None` when it
@@ -3004,7 +3003,7 @@ fn report(
 /// Under `RMLX_HOME`'s `tmp/`. The streams are a run artifact of a suite that
 /// names no output directory of its own, and a variable for one would be
 /// invisible configuration for a path the run already prints.
-fn write_round_stream(test: &str, prompt: &str, rounds: &[common::CapturedEvent]) {
+fn write_round_stream(test: &str, prompt: &str, rounds: &[CapturedEvent]) {
     let dir = rmlx_core::paths::tmp_dir();
     std::fs::create_dir_all(&dir).expect("create the run's tmp directory");
     let path = dir.join(format!("{test}.{prompt}.jsonl"));
@@ -3064,7 +3063,7 @@ fn emit_one_round_of_each_shape() {
 /// metadata whose target is in `BEHAVIOUR_SWITCH_TARGETS`.
 #[test]
 fn the_round_stream_recorder_keeps_a_round_and_charges_no_phase() {
-    let recorder = common::RoundStreamRecorder::new();
+    let recorder = RoundStreamRecorder::new();
     tracing::subscriber::with_default(std::sync::Arc::clone(&recorder), || {
         // The two questions a round loop asks before it decides what to do.
         assert!(!tracing::enabled!(target: "rmlx::spec::phase", tracing::Level::TRACE));
@@ -3074,20 +3073,29 @@ fn the_round_stream_recorder_keeps_a_round_and_charges_no_phase() {
         emit_one_round_of_each_shape();
     });
 
-    let rounds = common::round_events(&recorder.events());
+    let captured = recorder.events();
+    let rounds = round_events(&captured);
     assert_eq!(
         rounds.len(),
         3,
         "three rounds were emitted and the request-level line is not one: {rounds:?}"
     );
-    for round in &rounds {
-        for field in ["round", "accept", "num_draft"] {
-            assert!(
-                round.field(field).is_some(),
-                "a round the oracle can read carries {field}: {round:?}"
-            );
-        }
-    }
+    // The negative control has to be excluded for the reason `round_events`
+    // gives, not by luck: a request-level line that grew all three fields would
+    // otherwise quietly join the stream and this fixture would still read 3.
+    let control: Vec<&CapturedEvent> = captured
+        .iter()
+        .filter(|e| e.message == "a request-level line, not a round")
+        .collect();
+    assert_eq!(control.len(), 1, "the request-level line was captured");
+    assert!(
+        ROUND_EVENT_FIELDS
+            .iter()
+            .any(|f| control[0].field(f).is_none()),
+        "the line this fixture uses to show a non-round is excluded carries every one \
+         of {ROUND_EVENT_FIELDS:?}, so it no longer shows anything: {:?}",
+        control[0]
+    );
     assert_eq!(rounds[1].field("n_committed"), Some("3"));
     assert_eq!(rounds[2].field("projected_rows"), Some("3"));
 
@@ -3276,13 +3284,18 @@ fn run_gate(test: &str, pair: &Pair) {
     let mut judged = 0usize;
     for prompt in PROMPTS {
         let (block, spec, plain, margins, decided_by, rounds) = loaded.arms(prompt, device);
-        assert!(
-            !rounds.is_empty(),
-            "{test}/{}: the speculative arm reported no round, so the stream the \
-             comparison rests on is empty",
-            prompt.name
-        );
+        // Written before it is judged: an empty stream is the case worth having
+        // the file for.
         write_round_stream(test, prompt.name, &rounds);
+        assert!(
+            !rounds.is_empty() || spec.len() <= 1,
+            "{test}/{}: the speculative arm emitted {} tokens and reported no round. \
+             An arm that stopped on its seed closes no round and is the prompt; this \
+             one ran, so the stream the comparison rests on is empty because the loop \
+             stopped reporting one.",
+            prompt.name,
+            spec.len()
+        );
         // Every pair is judged at a block or not at all — the one it names, or
         // the one the serve layer resolves for a request that names none. The
         // readings below are all attributed to a block in the report line, and
